move some stuff around
