+++ src/mesh/forest/cell_center_offsets.hpp
@@ -25,7 +25,7 @@
 
 #include "basic_types.hpp"
 #include "defs.hpp"
-#include "mesh/logical_location.hpp"
+#include "mesh/forest/logical_location.hpp"
 #include "utils/bit_hacks.hpp"
 #include "utils/indexer.hpp"
