+++ src/mesh/forest/forest.cpp
@@ -29,7 +29,7 @@
 #include "mesh/forest/forest.hpp"
 #include "mesh/forest/relative_orientation.hpp"
 #include "mesh/forest/tree.hpp"
-#include "mesh/logical_location.hpp"
+#include "mesh/forest/logical_location.hpp"
 #include "utils/bit_hacks.hpp"
 #include "utils/indexer.hpp"
