+++ src/mesh/forest/forest.hpp
@@ -26,7 +26,7 @@
 #include "basic_types.hpp"
 #include "defs.hpp"
 #include "mesh/forest/tree.hpp"
-#include "mesh/logical_location.hpp"
+#include "mesh/forest/logical_location.hpp"
 #include "utils/bit_hacks.hpp"
 #include "utils/indexer.hpp"
