+++ src/mesh/forest/tree.hpp
@@ -27,7 +27,7 @@
 #include "defs.hpp"
 #include "mesh/forest/cell_center_offsets.hpp"
 #include "mesh/forest/relative_orientation.hpp"
-#include "mesh/logical_location.hpp"
+#include "mesh/forest/logical_location.hpp"
 #include "utils/bit_hacks.hpp"
 #include "utils/indexer.hpp"
